global: snap

+++ crates/vecdb/src/variants/lazy/vec/transform/halve.rs
@@ -5,7 +5,7 @@ use super::UnaryTransform;
 /// v -> v / 2
 pub struct Halve;
 
-impl<T: Div<i64, Output = T>> UnaryTransform<T> for Halve {
+impl<T: Div<usize, Output = T>> UnaryTransform<T> for Halve {
     #[inline(always)]
     fn apply(value: T) -> T {
         value / 2
